Remove crate_node_info metric
The metric wasn't useful for monitoring with prometheus.
It had a single static value and used the properties to include the
values for the node name and the node id. But these labels couldn't be
used in any way within prometheus.

In order to be useful these two labels would have to be applied to all
other metrics instead, so that they can be used for filtering.

But since prometheus is already node-aware by design and has to have
some form of "instance" information this additional node-id and
node-name information provides little value.